Update README.md
Added a link to a new crate I have recently released for genetic algorithms. Took some inspiration from this page and so wanted to be a apart of it also. The project is being developed pretty consistently so I think it would also benefit from the eyes of others, hopefully this list will give it some reach to those who can help build/make use of it!